rename: Pages 디렉토리 안에 파일이 과도하게 많은 이유로 디렉토리 구조 수정
