+++ src/pages/news/NewsPage.tsx
