+++ src/pages/teamroom/CoachBoardPageStyles.ts
